Reduce complexity of get redflag endpoints
[#162814389]
 On branch reduce-complexity
 Changes to be committed:
	modified:   .coverage
	modified:   api/app.py
	new file:   api/routes/get_incidents.py
	modified:   api/routes/incidents.py
