+++ api/routes/incidents.py
@@ -11,21 +11,13 @@ from api.models.incident import (
     RedFlag,
     incident_record_exists,
     red_flags,
-    get_incident_record,
-    get_all_incident_records,
-    get_incident_obj_by_id,
 )
 from api.helpers.validation import (
     validate_new_incident,
-    is_valid_id,
-    validate_edit_location,
-    request_data_required,
-    validate_comment,
 )
 
 from api.helpers.responses import (
     expected_new_incident_format,
-    delete_not_allowed,
 )
 
 
@@ -92,34 +84,3 @@ def new_red_flag():
         )
 
     return response
-
-
-@red_flags_bp.route("/red-flags", methods=["GET"])
-@token_required
-def get_all_red_flags():
-    return (
-        jsonify({"status": 200, "data": get_all_incident_records(red_flags)}),
-        200,
-    )
-
-
-@red_flags_bp.route("/red-flags/<red_flag_id>", methods=["GET"])
-@token_required
-def get_a_red_flag(red_flag_id):
-    try:
-        red_flag_id = int(red_flag_id)
-    except ValueError:
-        return (
-            jsonify(
-                {"status": 400, "error": "Red-flag id must be an integer"}
-            ),
-            400,
-        )
-
-    results = get_incident_record(red_flag_id, red_flags)
-    if results:
-        return jsonify({"status": 200, "data": [results]}), 200
-    return (
-        jsonify({"status": 404, "error": "Red-flag record does not exist"}),
-        404,
-    )
